WIP: Extract non-transport logic from TransportReplicationAction
Recent refactoring in TransportReplicationAction has moved all logic to clear units. We can take an extra step and extract those classes so they can be tested (and added to other tests) without requiring a (fake) transport layer and support.

This is a POC pending feedback. The usage of the new classes is shown in TransportReplicationAction2. Also, TransportIndexAction2 is an example implementation. Note how that class contains it's own IndexShardReference which is in charge of doing the actual work. This allows us in the future to extract class and use it to setup a unit test with multiple shards connected by the replication logic, but without the transport and other unneeded dependencies.