Reify publication (#21)
Prior to this change, the `publish()` method comprises a deeply nested 
collection of lambdas and anonymous classes which represent the notion of a
single publication attempt. In future we want to treat it as a first-class
concept so we can detect when it fails etc.

This change gives names to the anonymous lambdas and classes as a step towards
this.