Improve failure handling of PublicationTargets (#22)
Formerly, there was no explicit notion of failure of a PublicationTarget. This
change makes the state transitions of each target explicit and adds a FAILED 
state.

Additionally, formerly we sent ApplyCommit messages to all PublicationTargets
when a value was committed, but some of them may not even have received the
PublishRequest so were rejecting the ApplyCommit. This was hard to distinguish
from actual failures. This change alters the behaviour so that ApplyCommit are
only sent to nodes that have definitely received the corresponding
PublishRequest so that a rejection of ApplyCommit always represents a genuine
failure.

Finally, it's possible to be sure that this publication will fail if the
remaining non-failed targets do not form a quorum, and in this situation the
most sensible response is for the leader to stand down. Formerly we waited for
the publication to time out before doing so, but this change alters the
behaviour to do so explicitly.