Start the HeartbeatCollector before recording the local node (#26)
Previously, if the leader is a quorum all by itself then the call to
safeAddHeartbeatResponse(localNode) would immediately renew the leader's lease,
skipping the call to start() and suppressing sending heartbeats to all the
other nodes.  This breaks the assertion that, after stabilisation, all other
connected nodes are followers, because they become candidates if they do not
receive heartbeats.

After this change we call start() first so that heartbeats are always sent.