Be more specific than IllegalArgumentException (#25)
Today, we throw an overly-general IllegalArgumentException to reject
unacceptable messages in the consensus layer. We expect messages to be rejected
from time to time, if delayed or some other failure occurs, and it is not
useful to log the full stack trace leading to these exceptions.

This change introduces ConsensusMessageRejectedException for these cases, and
suppresses overly verbose logging when receiving such a message. It does not
include any of the extra machinery needed to pass these exceptions over the
wire.