Introduce the FutureExecutor (#29)
The FutureExecutor allows the legislator to schedule multiple tasks for future 
execution, rather than simply relying on handleWakeUp() to work out what needs
doing each time it is called.

The test implementation of FutureExecutor also adds a little extra variation in 
the scheduling of tasks. Apart from this, the rest of the implementation
remains unchanged: handleWakeUp() still works out what needs doing each time it
is called. This will be addressed in a subsequent change.