Make followers actively detect failures (#32)
Previously, each follower's has-the-leader-failed detector was passive, firing 
if it had not received a message from the leader sufficiently recently. This
may not detect asymmetric connectivity issues, and may not help in the 
situation where the follower has not yet joined the cluster fully.  Also the
leader only really cares about having a quorum of healthy followers, so would
not always take appropriate action if its connection with a single follower was
disrupted.

This change alters the behaviour to have the follower actively check that the 
leader is healthy, with a configurable number of retries allowed before
considering the leader to have failed. This also lets the follower check more 
things about the state of the leader before deciding on its health.