Add more detail to Legislator trace logs (#39)
It is sometimes challenging to work out the sequence of events that led to a
test failure, and a contributing factor is that scheduled tasks are identified
only by their scheduled time and not by any more useful identifier. This change
adds simple descriptions to these tasks to help untangle things a bit.