Allow fault-detection to fail publication targets (#42)
Today if a node fails then we attempt to remove it from the cluster state but
until this update is published we still unreasonably expect to be able to
communicate with it (e.g. during earlier publications). This change allows us
to fail faster, by failing the corresponding publication target as soon as a
failure is detected in any present and future publications. The node can only
be considered non-faulty again once a cluster state is published without it.