Track publications to completion (#49)
The MasterService ensures that only one publication is in progress at once.
However, the Legislator also currently tracks the set of active publications so
that they can be cancelled quickly by the fault-detection mechanism. Today we
stop tracking a publication once it is committed, but in future we must track
the publication until it either times out or receives a response from every
node.

This change modifies the Legislator's publication-tracking mechanism to track
publications until they complete, and also enforces that that there is only a
single active publication.