Retry pending tasks when publication completes (#50)
Today if a task submitted to the MasterService fails to result in a committed
update to the cluster state then we retry repeatedly until it succeeds (or the
master stands down).  A common reason for task failure is that there is already
a publication in flight, and in this case it makes more sense to wait for this
publication to complete before trying again. This change does exactly that.