Joining nodes should be considered non-faulty (#56)
Faulty nodes cease to be faulty when a cluster state is published in which they
have been removed. However, today, a faulty node may learn it is considered to
be faulty and send another join request to the master before its removal can be
published. In this case, the next publication will not remove it and so its
faultiness persists for longer than expected. This change fixes this by
removing joining nodes from the known-faulty set.