Improve support for iterated LegislatorTests (#58)
The Legislator tests are currently a little painful to run in an iterated
fashion (i.e. with `-Dtests.iters=BIGNUM`) due to the quantity of log output
produced. This change removes the TRACE-by-default annotations, and also
reduces the number of iterations in `runRandomly()`.

Verbose logging can be reinstated by running the tests with

    -Dtests.es.logger.level=TRACE

Fixes #52